+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 ### Added
 - `pip install htmd[dev]` to install development dependencies
 - Support Python 3.13 and 3.14
-- `htmd preview` will reload when posts and pages are modified
+- `htmd preview` will reload when posts are modified
 ### Fixed
 - Fix `htmd preview` when static directory is not the default